Update detect_shapes_resize.py
Add support for contour area calculation in pixels
Create --color and --area flags for deciding which properties the user wants displayed
Minor speed optimizations

NOTE: Area calculations may vary slightly from the "noresize" version of the calculation. This is due to  the image being resized before the contours are found, then the contours are multiplied by the resize ratio when the image returns to its original size. Because of this, the resized contours are slightly less detailed, leading to a slightly different area.